IMPORTANT: (#1) fix space-deletion following <Esc> press after match insertion (READ MSG)
I basically chose to temporarily remap <Enter> to insert the match
if the user has scrolled through the suggestions.

I've tried investigating the previous method, stripping it down
to the bear minimum code necessary and the bug would still
persist. Maybe in the future I'll find a way to actually
avoid messing with the user key-config, but for the moment
I wanna have something that actually works.